+++ CHANGELOG.rst
@@ -49,6 +49,9 @@ DrizzlePac (DEVELOPMENT)
 
 - Simplify logic for determining the chip ID for each source. [#200]
 
+- Added hlapipeline alignment code to package. [#216]
+
+- Added logging to the hlapipeline alignment code [#223]
 
 DrizzlePac v2.2.6 (02-Nov-2018)
 ===============================